requirements.txt
plotly
yfinance
Flask
bleak
requests
paho-mqtt
psutil